# docs/swagger.yaml
basePath: /
definitions:
  controllers.CustomerResponse:
    properties:
      id:
        type: string
    type: object
  controllers.HealthResponse:
    properties:
      services:
        additionalProperties:
          type: string
        example:
          mongodb: ok
          rabbitmq: ok
          redis: ok
        type: object
      status:
        example: ok
        type: string
    type: object
  controllers.MessageResponse:
    properties:
      message:
        type: string
    type: object
  controllers.OrderItemResponse:
    properties:
      id:
        type: string
      product_id:
        type: string
      product_name:
        type: string
      quantity:
        type: integer
      unit_price:
        type: integer
    type: object
  controllers.OrderResponse:
    properties:
      created_at:
        type: string
      customer_id:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/controllers.OrderItemResponse'
        type: array
      status:
        type: string
      total_amount:
        type: integer
      updated_at:
        type: string
    type: object
  controllers.ProductResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  controllers.UpdateStatusRequest:
    properties:
      status:
        type: string
    type: object
  dto.CreateOrderRequest:
    properties:
      customer_id:
        type: string
      items:
        items:
          $ref: '#/definitions/dto.OrderItem'
        type: array
    type: object
  dto.CreateProductRequest:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: integer
      stock:
        minimum: 0
        type: integer
    required:
    - name
    - price
    - stock
    type: object
  dto.OrderItem:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Order management API
  title: Challenge API
  version: "1.0"
paths:
  /api/v1/customers:
    post:
      description: Creates a new customer and returns its ID
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.CustomerResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a customer
      tags:
      - customers
  /api/v1/health:
    get:
      description: Checks the health of all dependent services
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.HealthResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controllers.HealthResponse'
      summary: Health check
      tags:
      - health
  /api/v1/orders:
    post:
      consumes:
      - application/json
      description: Creates a new order with stock deduction and idempotency support
      parameters:
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      - description: Order data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create an order
      tags:
      - orders
  /api/v1/orders/{id}:
    get:
      description: Returns a single order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get order by ID
      tags:
      - orders
  /api/v1/orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: Updates the status of an existing order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update order status
      tags:
      - orders
  /api/v1/products:
    get:
      description: Returns all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.ProductResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List all products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Creates a new product
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a product
      tags:
      - products
swagger: "2.0"